// src/pages/order-domain/components/enums.ts
export enum StatusOrderDomainEnum {
    ORDER_STATUS_SEO_CREATE = 'SEO_CREATE_ORDER',
    ORDER_STATUS_CREATE = 'CREATE_ORDER',
    ORDER_STATUS_DONE = 'DONE_ORDER',
    ORDER_STATUS_CANCEL = 'CANCEL_ORDER',
    ORDER_STATUS_ASYNC = 'ASYNC_ORDER',
    UN_SUCCESS = 'UN_SUCCESS',
    // ORDER_STATUS_APPROVED = 'APPROVED_ORDER',
}

export enum STATUS_DOMAIN_IN_CART {
    NEW_DOMAIN_ORDER = 'Đề xuất mua',
    BUY_SUCCESS = 'Mua thành công',
    ERROR_CONNECT = 'Lỗi kết nối',
    NOT_AVAILABLE = 'Domain không có sẵn',
    WALLET_OUT_OF_MONEY = 'Ví tài khoản không đủ tiền',
    CANCEL_DOMAIN = 'Hủy domain',
    SUBMIT_REGISTRATION = 'Gửi đăng ký thành công, vui lòng đợi hệ thống xử lý, số tiền đóng băng $ 3,00',
}
  